Split server-payments.test.ts into focused files
The monolithic test/lib/server-payments.test.ts (1654 lines) was grandfathered
into the noExcessiveLinesPerFile biome exclusion list. Split it into a
test/lib/server-payments/ folder — one file per concern (success errors,
idempotent replay, cancel, checkout/purchase, confirmation) — and drop its
exclusion entry.

Splitting surfaced duplication that jscpd could not scan inside the monolith.
Fixed it by reusing the existing shared stripe stub helpers
(stubRetrieveCheckoutSession / stubRefundPayment) instead of hand-rolled
inline session objects, extracting a shared findKeptPlaceholder helper (also
adopted in the multi-ticket webhook test), and adding small local helpers for
the repeated signed-session, cancelled-session, checkout-intent-capture and
replayed-save shapes. jscpd is back to 0 clones.
