+++ biome.json
@@ -156,7 +156,6 @@
         "src/shared/db/migrations/schema.ts",
         "src/ui/templates/public/reservations.tsx",
         "test/lib/server-listing-parents.test.ts",
-        "test/lib/server-payments.test.ts",
         "test/lib/square.test.ts",
         "test/lib/stripe.test.ts"
       ],
